src/dcs/outliers/requirements.txt
opencv-python-headless
numpy
pandas
scikit-learn
plotly
dash
pillow
